extract WebsocketServerConnector
