+++ packages/dht/src/connection/connectivityRequestHandler.ts
@@ -9,7 +9,7 @@ import { NatType } from './ConnectionManager'
 import { CONNECTIVITY_CHECKER_SERVICE_ID, connectAsync } from './connectivityChecker'
 import { IConnection } from './IConnection'
 import { WebsocketServerConnection } from './websocket/WebsocketServerConnection'
-import { connectivityMethodToWebsocketUrl } from './websocket/WebsocketConnector'
+import { connectivityMethodToWebsocketUrl } from './websocket/WebsocketClientConnector'
 import { LOCAL_PROTOCOL_VERSION } from '../helpers/version'
 import { GeoIpLocator } from '@streamr/geoip-location'
